+++ src/test/java/PersonTest.java
@@ -5,7 +5,7 @@ import java.awt.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
-public class PersonTest {
+class PersonTest {
 
     @Test
     void personHasFirstName() {
